🧹 refactor: centralize utilities and fix frontend timer leaks
- Centralized duplicate `maskKey`, `escapeHtml`, and `generateOTP` functions into `Backend/utils/strings.js` and `Backend/utils/security.js`.
- Moved cookie extraction and header building logic to `Backend/utils/cookies.js`.
- Standardized backend responses to use `res.json()` and removed unused dependencies from `user.controller.js`.
- Fixed a syntax error in `Categories.jsx` where style properties were incorrectly nested in a className.
- Implemented robust `setTimeout` and `setInterval` cleanup using `useRef` in `Login.jsx`, `ChatRaj.jsx`, `Register.jsx`, and `VimCodeEditor.jsx` to prevent memory leaks.
- Centralized rate limiters in `Backend/middleware/rateLimiter.js`.
