+++ Backend/middleware/rateLimiter.js
@@ -17,4 +17,30 @@ export const authLimiter = rateLimit({
   legacyHeaders: false,
 });
 
-export default { sensitiveLimiter, authLimiter };
+export const leaderboardLimiter = rateLimit({
+  windowMs: 15 * 60 * 1000, // 15 minutes
+  max: 100, // limit each IP to 100 leaderboard requests per windowMs
+});
+
+export const usersLimiter = rateLimit({
+  windowMs: 15 * 60 * 1000, // 15 minutes
+  max: 100, // limit each IP to 100 user list requests per windowMs
+});
+
+// Slightly higher limit for registration to reduce false positives
+// during legitimate user signups (e.g., when users retry due to client-side
+// behavior).
+export const registrationLimiter = rateLimit({
+  windowMs: 15 * 60 * 1000,
+  max: 12,
+  standardHeaders: true,
+  legacyHeaders: false,
+});
+
+export default {
+  sensitiveLimiter,
+  authLimiter,
+  leaderboardLimiter,
+  usersLimiter,
+  registrationLimiter
+};
